+++ internal/api/server.go
@@ -11,7 +11,6 @@ import (
 	"net/http"
 	"os"
 	"path"
-	"regexp"
 	"runtime/debug"
 	"strconv"
 	"strings"
@@ -45,10 +44,6 @@ const (
 	gqlEndpoint        = "/graphql"
 	playgroundEndpoint = "/playground"
 )
-
-var version string
-var buildstamp string
-var githash string
 
 var uiBox = ui.UIBox
 var loginUIBox = ui.LoginUIBox
@@ -270,7 +265,7 @@ func Start() error {
 		TLSNextProto: make(map[string]func(*http.Server, *tls.Conn, http.Handler)),
 	}
 
-	printVersion()
+	config.PrintVersion()
 	go printLatestVersion(context.TODO())
 	logger.Infof("stash is listening on " + address)
 	if tlsConfig != nil {
@@ -388,49 +383,6 @@ func customLocalesHandler(c *config.Instance) func(w http.ResponseWriter, r *htt
 		w.Header().Set("Content-Type", "application/json")
 		utils.ServeStaticContent(w, r, buffer.Bytes())
 	}
-}
-
-func printVersion() {
-	var versionString string
-	switch {
-	case version != "":
-		if githash != "" && !IsDevelop() {
-			versionString = version + " (" + githash + ")"
-		} else {
-			versionString = version
-		}
-	case githash != "":
-		versionString = githash
-	default:
-		versionString = "unknown"
-	}
-	if config.IsOfficialBuild() {
-		versionString += " - Official Build"
-	} else {
-		versionString += " - Unofficial Build"
-	}
-	if buildstamp != "" {
-		versionString += " - " + buildstamp
-	}
-	logger.Infof("stash version: %s\n", versionString)
-}
-
-func GetVersion() (string, string, string) {
-	return version, githash, buildstamp
-}
-
-func IsDevelop() bool {
-	if githash == "" {
-		return false
-	}
-
-	// if the version is suffixed with -x-xxxx, then we are running a development build
-	develop := false
-	re := regexp.MustCompile(`-\d+-g\w+$`)
-	if re.MatchString(version) {
-		develop = true
-	}
-	return develop
 }
 
 func makeTLSConfig(c *config.Instance) (*tls.Config, error) {
