+++ internal/manager/config/init.go
@@ -25,6 +25,7 @@ type flagStruct struct {
 	cpuProfilePath string
 	nobrowser      bool
 	helpFlag       bool
+	versionFlag    bool
 }
 
 func GetInstance() *Instance {
@@ -44,6 +45,11 @@ func Initialize() (*Instance, error) {
 
 		if flags.helpFlag {
 			pflag.Usage()
+			os.Exit(0)
+		}
+
+		if flags.versionFlag {
+			fmt.Printf(GetVersionString() + "\n")
 			os.Exit(0)
 		}
 
@@ -134,6 +140,7 @@ func initFlags() flagStruct {
 	pflag.StringVar(&flags.cpuProfilePath, "cpuprofile", "", "write cpu profile to file")
 	pflag.BoolVar(&flags.nobrowser, "nobrowser", false, "Don't open a browser window after launch")
 	pflag.BoolVarP(&flags.helpFlag, "help", "h", false, "show this help text and exit")
+	pflag.BoolVarP(&flags.versionFlag, "version", "v", false, "show version number and exit")
 
 	pflag.Parse()
