+++ internal/manager/config/version.go
@@ -1,0 +1,63 @@
+package config
+
+import (
+	"regexp"
+
+	"github.com/stashapp/stash/pkg/logger"
+)
+
+var version string
+var buildstamp string
+var githash string
+var officialBuild string
+
+func GetVersion() (string, string, string) {
+	return version, githash, buildstamp
+}
+
+func GetVersionString() string {
+	var versionString string
+	switch {
+	case version != "":
+		if githash != "" && !IsDevelop() {
+			versionString = version + " (" + githash + ")"
+		} else {
+			versionString = version
+		}
+	case githash != "":
+		versionString = githash
+	default:
+		versionString = "unknown"
+	}
+	if IsOfficialBuild() {
+		versionString += " - Official Build"
+	} else {
+		versionString += " - Unofficial Build"
+	}
+	if buildstamp != "" {
+		versionString += " - " + buildstamp
+	}
+	return versionString
+}
+
+func PrintVersion() {
+	logger.Infof("stash version: %s\n", GetVersionString())
+}
+
+func IsOfficialBuild() bool {
+	return officialBuild == "true"
+}
+
+func IsDevelop() bool {
+	if githash == "" {
+		return false
+	}
+
+	// if the version is suffixed with -x-xxxx, then we are running a development build
+	develop := false
+	re := regexp.MustCompile(`-\d+-g\w+$`)
+	if re.MatchString(version) {
+		develop = true
+	}
+	return develop
+}
